+++ mininterface/web_interface/__init__.py
@@ -8,7 +8,7 @@ from ..mininterface.adaptor import MinAdaptor
 
 from ..form_dict import EnvClass
 
-from ..options import UiOptions
+from ..settings import UiSettings
 
 from ..textual_interface import TextualInterface
 from .child_adaptor import SerializedChildAdaptor
@@ -30,7 +30,7 @@ class WebInterface(TextualInterface):
 
     def __init__(self,
                  title: str = "",
-                 options: Optional[UiOptions] = None,
+                 settings: Optional[UiSettings] = None,
                  _env: EnvClass | SimpleNamespace | None = None,
                  cmd: Optional[str] = None, port=64646, **kwargs):
         # TODO
@@ -62,15 +62,15 @@ class WebInterface(TextualInterface):
         #     m.form({"hello": 1})  # the app ends here
         #     m.form({"hello": 2})  # we never get here
 
-        super().__init__(title, options, _env, need_atty=False, **kwargs)
+        super().__init__(title, settings, _env, need_atty=False, **kwargs)
         match os.environ.get("MININTERFACE_ENFORCED_WEB"):
             case '_web-child-serialized':
-                self._adaptor = SerializedChildAdaptor(self, options)
+                self._adaptor = SerializedChildAdaptor(self, settings)
                 return
             case '_web-parent':
                 envir = os.environ.copy()
                 envir["MININTERFACE_ENFORCED_WEB"] = '_web-child-serialized'
-                self._adaptor = WebParentAdaptor(self, options, environ=envir)
+                self._adaptor = WebParentAdaptor(self, settings, environ=envir)
                 self._adaptor.disconnect()
                 quit()
             case _: